requirements.txt
# Web Framework
fastapi
uvicorn
python-multipart

# AI & OCR
# Note: paddlepaddle is the CPU version
paddlepaddle
paddleocr
shapely
opencv-python

# Databases
# MongoDB (Async driver)
motor
# PostgreSQL (SQLAlchemy + Driver)
sqlalchemy
psycopg2-binary

# Utilities
python-dotenv
pydantic
pydantic-settings